Response to Issue#32: Let TimeStamp be integer
Timestamp can be left as an integer. And it can be interpreted as `datetime.datetime()` object when required to be printed. This is because keeping it as an integer it can be converted into other timestamp formats like the SAS Timestamp format that starts with January 1, 1960, unlike the UNIX Time.  In this way, this parser can be used with other software also.